refactor (providers): move context to separate folder
